+++ demo/index.js
@@ -2,6 +2,9 @@
 /* global ReactQuill */
 'use strict';
 
+if (typeof React !== 'object') alert('React not found. Did you run "npm install"?');
+if (typeof ReactQuill !== 'function') alert('ReactQuill not found. Did you run "make build"?')
+
 var Editor = React.createClass({
 
 	getInitialState: function() {
@@ -9,14 +12,14 @@ var Editor = React.createClass({
 			theme: 'snow',
 			enabled: true,
 			readOnly: false,
-			value: '<h1>It works!</h1>',
+			value: '',
 			events: []
 		};
 	},
 
 	formatRange: function(range) {
 		return range
-			? [range.start, range.end].join(',')
+			? [range.index, range.index + range.length].join(',')
 			: 'none';
 	},