+++ src/index.js
@@ -1,8 +1,10 @@
 /*
-React-Quill v0.4.1
+React-Quill v1.0.0
 https://github.com/zenoamaro/react-quill
 */
+var Quill = require('quill');
+
 module.exports = require('./component');
+module.exports.Quill = Quill;
 module.exports.Mixin = require('./mixin');
 module.exports.Toolbar = require('./toolbar');
-module.exports.Quill = require('quill');